// src/app/admin/transactions/models/OpenVASP.Messaging.Messages.Entities/natural-identification-type.enum.ts
export enum NaturalIdentificationType {
  PassportNumber = 'PassportNumber',
  NationalIdentityNumber = 'NationalIdentityNumber',
  SocialSecurityNumber = 'SocialSecurityNumber',
  TaxIdentificationNumber = 'TaxIdentificationNumber',
  AlienRegistrationNumber = 'AlienRegistrationNumber',
  DriversLicenseNumber = 'DriversLicenseNumber',
  Other = 'Other',
}